Ryddet i koden til modul user-token-exchange.
