--- tokendings-exchange/src/main/kotlin/no/nav/tms/token/support/tokendings/exchange/TokenXEnvironment.kt
@@ -1,23 +1,0 @@
-package no.nav.tms.token.support.tokendings.exchange
-
-// Proxy for System environment which allows for mocking or overwriting default env
-object TokenXEnvironment {
-    private val baseEnv = System.getenv()
-
-    private val env = mutableMapOf<String, String>()
-
-    init {
-        env.putAll(baseEnv)
-    }
-
-    fun get(name: String) = env[name]
-
-    fun extend(envMap: Map<String, String>) {
-        env.putAll(envMap)
-    }
-
-    fun reset() {
-        env.clear()
-        env.putAll(baseEnv)
-    }
-}
