--- tokendings-exchange/src/main/kotlin/no/nav/tms/token/support/tokendings/exchange/TokenXHeader.kt
@@ -1,5 +1,0 @@
-package no.nav.tms.token.support.tokendings.exchange
-
-object TokenXHeader {
-    const val Authorization = "token-x-authorization"
-}
