--- tokendings-exchange/src/main/kotlin/no/nav/tms/token/support/tokendings/exchange/TokendingsService.kt
@@ -1,5 +1,0 @@
-package no.nav.tms.token.support.tokendings.exchange
-
-interface TokendingsService {
-    suspend fun exchangeToken(token: String, targetApp: String): String
-}
