--- tokendings-exchange/src/main/kotlin/no/nav/tms/token/support/tokendings/exchange/config/HttpClientBuilder.kt
@@ -1,22 +1,0 @@
-package no.nav.tms.token.support.tokendings.exchange.config
-
-import com.fasterxml.jackson.databind.DeserializationFeature
-import io.ktor.client.*
-import io.ktor.client.engine.apache.*
-import io.ktor.client.plugins.*
-import io.ktor.client.plugins.contentnegotiation.*
-import io.ktor.serialization.jackson.*
-
-internal object HttpClientBuilder {
-    internal fun buildHttpClient(): HttpClient {
-        return HttpClient(Apache) {
-            install(ContentNegotiation) {
-                jackson {
-                    configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false)
-                }
-            }
-            install(HttpTimeout)
-        }
-    }
-}
-
