--- tokendings-exchange/src/main/kotlin/no/nav/tms/token/support/tokendings/exchange/config/cache/AccessTokenKey.kt
@@ -1,7 +1,0 @@
-package no.nav.tms.token.support.tokendings.exchange.config.cache
-
-internal data class AccessTokenKey(
-        val subject: String,
-        val securityLevel: String,
-        val targetApp: String
-)
