--- user-token-exchange/src/main/kotlin/no/nav/tms/token/support/tokendings/exchange/config/TokendingsContext.kt
@@ -1,20 +1,0 @@
-package no.nav.tms.token.support.tokendings.exchange.config
-
-import io.ktor.client.*
-import kotlinx.coroutines.runBlocking
-import no.nav.tms.token.support.tokendings.exchange.consumer.TokendingsConsumer
-
-
-internal class TokendingsContext {
-    private val httpClient = HttpClientBuilder.buildHttpClient()
-
-    internal val environment = Environment()
-
-    internal val metadata = fetchMetadata(httpClient, environment.tokenxWellKnownUrl)
-
-    internal val tokendingsConsumer = TokendingsConsumer(httpClient, metadata.tokenEndpoint)
-}
-
-private fun fetchMetadata(httpClient: HttpClient, tokendingsUrl: String) = runBlocking {
-    httpClient.getTokendingsConfigurationMetadata(tokendingsUrl)
-}
