--- user-token-exchange/src/main/kotlin/no/nav/tms/token/support/tokendings/exchange/config/httpClient.kt
@@ -1,17 +1,0 @@
-package no.nav.tms.token.support.tokendings.exchange.config
-
-import io.ktor.client.*
-import io.ktor.client.call.*
-import io.ktor.client.request.*
-import io.ktor.http.*
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
-
-internal suspend fun HttpClient.getTokendingsConfigurationMetadata(url: String)
-        : TokendingsConfigurationMetadata = withContext(Dispatchers.IO) {
-    request {
-        method = HttpMethod.Get
-        url(url)
-        accept(ContentType.Application.Json)
-    }.body()
-}
