+++ user-token-exchange/src/main/kotlin/no/nav/tms/token/support/user/token/exchange/UserTokenExchangeService.kt
@@ -1,0 +1,10 @@
+package no.nav.tms.token.support.user.token.exchange
+
+interface UserTokenExchangeService {
+    suspend fun exchangeToken(token: String, targetApp: String): String
+}
+
+class UserTokenExchangeException(cause: Exception, target: String):
+    RuntimeException("Feil ved veksling av token for å nå app [$target]", cause)
+
+
