+++ user-token-exchange/src/main/kotlin/no/nav/tms/token/support/user/token/exchange/service/UserTokenExchangeResponse.kt
@@ -1,0 +1,5 @@
+package no.nav.tms.token.support.user.token.exchange.service
+
+import com.fasterxml.jackson.annotation.JsonAlias
+
+
