--- user-token-exchange/src/test/kotlin/no/nav/tms/token/support/tokendings/exchange/TokendingsResponseObjectMother.kt
@@ -1,12 +1,0 @@
-package no.nav.tms.token.support.tokendings.exchange
-
-import no.nav.tms.token.support.tokendings.exchange.consumer.TokendingsTokenResponse
-
-internal object TokendingsResponseObjectMother {
-    fun createTokendingsResponse(accessToken: String, expiresIn: Int = 300) = TokendingsTokenResponse (
-            accessToken = accessToken,
-            tokenType = "",
-            issuedTokenType = "",
-            expiresIn = expiresIn
-    )
-}
